Moved update functions back into `config.py` because of import issues.
